IFXContactCreateEdit is now a shell for IFXEmailContact, IFXPhoneContact and IFXFullContact which have the actual forms.
