+++ src/components/contact/IFXContactCreateEdit.vue
@@ -1,25 +1,22 @@
 <script>
-import { mapActions } from 'vuex'
-import IFXItemCreateEditMixin from '@/components/item/IFXItemCreateEditMixin'
-import IFXContactMixin from '@/components/contact/IFXContactMixin'
-
 export default {
   name: 'IFXContactCreateEdit',
-  mixins: [IFXContactMixin, IFXItemCreateEditMixin],
+  props: {
+    contactType: {
+      required: false,
+      default: null
+    },
+    id: {
+      required: false,
+    }
+  },
   data() {
     return {
       allUsers: [],
+      localContactType: this.contactType
     }
   },
   methods: {
-    ...mapActions(['showMessage']),
-    async init() {
-      this.item = await this.getItem()
-      this.cachedItem = JSON.parse(JSON.stringify(this.item))
-      // TODO: only get chunks
-      this.allUsers = await this.$api.user.getList()
-        .catch(err => this.showMessage(err))
-    },
     updateContactUsers(users) {
       this.$set(this.item.users, ...users)
     },
@@ -37,82 +34,42 @@ export default {
       <template #content>{{description}}</template>
     </IFXPageHeader>
     <v-container>
-      <v-form v-if='!isLoading' v-model='isValid'>
-        <v-row>
-          <v-col>
-            <v-text-field
-              v-model='item.name'
-              label='Name'
-              data-cy='name'
-              :rules='formRules.generic'
-              :error-messages='errors.name'
-              required
-            ></v-text-field>
-          </v-col>
-          <v-col>
-            <v-text-field
-              v-model='item.detail'
-              label='Email'
-              data-cy='email'
-              :rules='formRules.email'
-              :error-messages='errors.detail'
-              required
-            ></v-text-field>
-          </v-col>
-        </v-row>
-        <v-row>
-          <v-col>
-            <v-select
-              v-model='item.type'
-              label='Type'
-              data-cy='type'
-              :items='apiRef.types'
-              :rules='formRules.generic'
-              :error-messages='errors.type'
-              required
-            ></v-select>
-          </v-col>
-          <v-col>
-            <v-text-field
-              v-model='item.phone'
-              label='Phone'
-              type='number'
-              data-cy='phone'
-              :rules='formRules.phone'
+      <v-row>
+        <v-col>
+          Select a Contact type
+          <v-radio-group
+            v-model="localContactType"
+            row
+          >
+            <v-radio
+              label="Email"
+              value="Email"
             >
-            </v-text-field>
-          </v-col>
-        </v-row>
-        <v-row>
-          <v-col>
-            <v-text-field
-              v-model='item.address'
-              label='Address'
-              data-cy='address'
-            ></v-text-field>
-          </v-col>
-        </v-row>
-        <!-- TODO: Add reverse search for users and org contacts associated with this contact -->
-        <!-- <span v-if='isEditing'>
-          <v-row>
-            <v-col>
-              <IFXItemSelectList title='Users' :items='contact.users' disabled v-slot="{ item }">
-                <IFXUserEdit :item='item'/>
-              </IFXItemSelectList>
-            </v-col>
-            <v-col>
-              <IFXItemSelectList title='Organization Contacts' :items='contact.contacts' disabled v-slot="{ item }">
-                <IFXOrganizationContactEdit :item='item'/>
-              </IFXItemSelectList>
-            </v-col>
-          </v-row>
-        </span> -->
-        <v-row>
-          <v-col>
-            <IFXButton :disabled='!isSubmittable' btnType='submit' @action='submit' />
-          </v-col>
-        </v-row>
-      </v-form>
+            </v-radio>
+            <v-radio
+              label="Phone"
+              value="Phone"
+            >
+            </v-radio>
+            <v-radio
+              label="Full"
+              value="Full"
+            >
+            </v-radio>
+          </v-radio-group>
+        </v-col>
+      </v-row>
+      <v-row>
+        <v-col v-if="localContactType === 'Email'">
+          <IFXEmailContact :id="id"/>
+        </v-col>
+        <v-col v-else-if="localContactType === 'Phone'">
+          <IFXPhoneContact :id="id"/>
+        </v-col>
+        <v-col v-else>
+          <IFXFullContact :id="id"/>
+        </v-col>
+      </v-row>
     </v-container>
   </v-container>
 </template>
