+++ src/components/contact/IFXEmailContact.vue
@@ -1,0 +1,46 @@
+<script>
+import { mapActions } from 'vuex'
+import IFXItemCreateEditMixin from '@/components/item/IFXItemCreateEditMixin'
+import IFXContactMixin from '@/components/contact/IFXContactMixin'
+
+export default {
+  name: 'IFXEmailContact',
+  mixins: [IFXContactMixin, IFXItemCreateEditMixin],
+  data() {
+    return {}
+  },
+  methods: {
+    ...mapActions(['showMessage']),
+    async init() {
+      this.item = await this.getItem()
+      this.cachedItem = JSON.parse(JSON.stringify(this.item))
+    },
+  },
+  mounted() {
+    this.item.type = 'Email'
+  }
+}
+</script>
+<template>
+  <v-container>
+    <v-form v-if='!isLoading' v-model='isValid'>
+      <v-row>
+        <v-col>
+          <v-text-field
+            v-model='item.detail'
+            label='Email'
+            data-cy='email'
+            :rules='formRules.email'
+            :error-messages='errors.detail'
+            required
+          ></v-text-field>
+        </v-col>
+      </v-row>
+      <v-row>
+        <v-col>
+          <IFXButton :disabled='!isSubmittable' btnType='submit' @action='submit' />
+        </v-col>
+      </v-row>
+    </v-form>
+  </v-container>
+</template>
